ble_detector: recognize 8 more device types (2.1.9)
Adds Tile, Beats, Apple host (iPhone/Watch/iPad/Mac Nearby Info), Xiaomi,
Galaxy Buds, Garmin, Fitbit and GoPro to the BLE detector, via service
UUIDs, company IDs, Apple continuity model bytes and name hints. The kind
bitfield widens from 16 to 32 bits (DetectorMatch.kinds/signatures and the
per-record kinds) since there are now 19 kinds; all consumers already
iterate DetectorKindCount and mask with DETECTOR_BIT, so no other change
was needed. Host parser test (tests/host/test_ble_detector_parse.c)
extended and passes.

Detector expansion developed with DeepSeek assistance in-tree; verified,
built and shipped here.

+++ applications/main/ble_detector/ble_detector_app.c
@@ -28,7 +28,7 @@ typedef struct {
     uint8_t addr_type;
     int8_t rssi;
     uint32_t last_seen;
-    uint16_t kinds;
+    uint32_t kinds;
     char name[32];
 } DetectorRecord;
